requirements.txt
google_api_python_client==2.86.0
pandas==2.0.0
PyYAML==6.0
PyYAML==6.0
Requests==2.28.2
